Update README.md
I'm working through the C# track, and had a bit of trouble understanding what I needed to do for this exercise.  I was able to figure it out by looking at the tests, but I thought there might be a clearer way to explain things in the Readme.  I really just removed the reference to mnemonics, which was interesting but made me confused about what I needed to do.  Thanks for your work!